Improve org structure editing
C-RET, M-RET, and C-M-RET respectively insert a heading at the
end of the subtree, a heading after the current line, or a
subheading after the current line. org-insert-heading was patched
so that the latter two bindings respect property drawers, inserting
after them if one is present. With org-insert-heading-respect-content
disabled, the latter two bindings will also insert above if used at
the start of a line.