Patch +evil-join-a
Other parts of Doom require smartparens, which populates
doom-point-in-comment-functions with sp-point-in-comment, which
prevents doom-point-in-comment-p from returning nil (returned by
syntax-ppss when point is on the first comment delimiter). This
prevents evil-join-a from exiting early, without removing delimiters
when lines are joined. NOTE This happens regardless of whether you've
disabled smartparens-global-mode and removed any smartparens flags
from modules like :default.